requirements.txt
git+https://github.com/just-ai/mlp-python-sdk.git@dev
python-dotenv
pydantic
openai
distutils